+++ CHANGELOG.md
@@ -6,6 +6,8 @@ All notable changes to this project will be documented in this file.
 ## [Unreleased] — 
 
 ### Added
+- **CLI command `report tree`** — `--from` and `--to` tree view of Counterparty -> Category -> Operations.
+- **CLI command `operation update`** — `--description \ -d`, `--counterparty \ -c`and  `--category \ -g`.
 - **CLI command `search`** — `--counterparty \ -c` and `--category \ -g` flags and wire them through to the search layer. search `-c|-g None` returns the undefined categories and counterparties, respectively.
 - **CLI command `report counterparty`** — support `--from` and `--to`.
 - **CLI command `report category`** — support `--from` and `--to`.